requirements.txt
customtkinter>=5.2.0
pygments>=2.15.0
pyserial>=3.5
chardet>=5.1.0
intelhex>=2.3.0
pyserial>=3.5

customtkinter>=5.2.2
Pillow>=10.0.0
pyserial>=3.5
intelhex>=2.3.0